+++ src/BigDecimal.php
@@ -28,10 +28,8 @@ use function rtrim;
 use function str_repeat;
 use function strlen;
 use function substr;
-use function trigger_error;
 use function unpack;
 
-use const E_USER_DEPRECATED;
 use const PHP_INT_SIZE;
 
 /**
@@ -822,21 +820,6 @@ final readonly class BigDecimal extends BigNumber
 
         /** @var positive-int */
         return ($this->value[0] === '-') ? $length - 1 : $length;
-    }
-
-    /**
-     * Returns whether this decimal number has a non-zero fractional part.
-     *
-     * @deprecated Will be removed in 0.17. Use `! $number->getFractionalPart()->isZero()` instead.
-     */
-    public function hasNonZeroFractionalPart(): bool
-    {
-        trigger_error(
-            'BigDecimal::hasNonZeroFractionalPart() is deprecated and will be removed in 0.17. Use `! $number->getFractionalPart()->isZero()` instead.',
-            E_USER_DEPRECATED,
-        );
-
-        return ! $this->getFractionalPart()->isZero();
     }
 
     /**